Move unintentionally-exported code to private modules
