+++ docs/api/viz.md
@@ -4,6 +4,6 @@ The top-level namespace accessible after `import lonboard`.
 
 Layers are also accessible in the top-level namespace, but are documented in the `layers/` folder.
 
-::: lonboard.viz.viz
+::: lonboard.viz
     options:
       show_root_full_path: false
