+++ lonboard/__init__.py
@@ -1,3 +1,5 @@
+from . import colormap, traits
+from ._layer import PathLayer, ScatterplotLayer, SolidPolygonLayer
+from ._map import Map
 from ._version import __version__
-from .layer import PathLayer, ScatterplotLayer, SolidPolygonLayer
-from .viz import viz
+from ._viz import viz
