+++ lonboard/_geoarrow/ops/centroid.py
@@ -8,7 +8,7 @@ from typing import Optional
 import numpy as np
 import pyarrow as pa
 
-from lonboard.constants import EXTENSION_NAME
+from lonboard._constants import EXTENSION_NAME
 
 
 @dataclass
