+++ lonboard/_serialization.py
